Added GitIgnore and some files for fun
AsciiReader prints to a file java code to print a given hardcoded ascii.
Takes an ascii file as input.

Kattio is a input / output program for java that is very reliant and
good, sponsored by KTH.

rconMaker makes your configs available in a rented server (where rcon
password is required). Currently in Alpha stage, version 4.0. Takes a
config file as input and gives a new file for the user that is rcon
friendly